Factor out BundleDetails and AssetDetails
